Fixed bug in implementation of threading
The existence of a cache (from a previous commit's feature) hid a bug in
the implementation of threading!  The execution of self.clear_screen()
immediately after that branch hid a number of useful error messages,
which made debugging quite annoying. Fixed bug in threading
implementation with some ```python @staticmethod``` black magic.